SAM2/requirements.txt
torch>=2.3.1
torchvision>=0.18.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
pillow>=9.4.0
matplotlib>=3.9.0
opencv
scikit-image
seaborn
pyradiomics
umap-learn
hdbscan